Note section part 4 is done

+++ app/Http/Controllers/NoteController.php
@@ -12,8 +12,10 @@ class NoteController extends Controller
      */
     public function index()
     {
-        //
-        $notes = Note::query()->orderBy('created_at', 'desc')->paginate();
+        $notes = Note::query()
+            ->where('user_id', request()->user()->id)
+            ->orderBy('created_at', 'desc')
+            ->paginate();
         return view('note.index', ['notes' => $notes]);
     }
 
@@ -22,7 +24,6 @@ class NoteController extends Controller
      */
     public function create()
     {
-        //
         return view('note.create');
     }
 
@@ -31,15 +32,14 @@ class NoteController extends Controller
      */
     public function store(Request $request)
     {
-        //
         $data = $request->validate([
             'note' => ['required', 'string']
         ]);
 
-        $data['user_id'] = 1;
+        $data['user_id'] = $request->user()->id;
         $note = Note::create($data);
 
-        return to_route('note.show', $note)->with('message', 'Note was created');
+        return to_route('note.show', $note)->with('message', 'Note was create');
     }
 
     /**
@@ -47,7 +47,9 @@ class NoteController extends Controller
      */
     public function show(Note $note)
     {
-        //
+        if ($note->user_id !== request()->user()->id) {
+            abort(403);
+        }
         return view('note.show', ['note' => $note]);
     }
 
@@ -56,7 +58,9 @@ class NoteController extends Controller
      */
     public function edit(Note $note)
     {
-        //
+        if ($note->user_id !== request()->user()->id) {
+            abort(403);
+        }
         return view('note.edit', ['note' => $note]);
     }
 
@@ -65,7 +69,9 @@ class NoteController extends Controller
      */
     public function update(Request $request, Note $note)
     {
-        //
+        if ($note->user_id !== request()->user()->id) {
+            abort(403);
+        }
         $data = $request->validate([
             'note' => ['required', 'string']
         ]);
@@ -80,7 +86,9 @@ class NoteController extends Controller
      */
     public function destroy(Note $note)
     {
-        //
+        if ($note->user_id !== request()->user()->id) {
+            abort(403);
+        }
         $note->delete();
 
         return to_route('note.index')->with('message', 'Note was deleted');
